common/pyproject.toml
[tool.poetry]
name = "common"
description = "This is how python package should look like!"
version = "0.1.0"
license = "MIT"

authors = []

readme = "README.md"

repository = "https://github.com/aljeshishe/dashboards"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.26.27"
requests = "^2.28.1"
munch = "^2.5.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sls2env = 'common.sls2env:main'
